+++ acme/services/ActionManager.py
@@ -100,7 +100,6 @@ class ActionManager(object):
 
 		# Get actions. Remember, these are NOT <action> resources
 		actions = CSE.storage.searchActionsForSubject(_ri)		
-		L.logWarn(actions)
 		# sort by action priority
 		actions = sorted(actions, key = lambda x: x['apy'] if x['apy'] is not None else sys.maxsize)